Refactor GenericUnitTest into AssemblyOperations, CacheManager and ConfigManager
