+++ src/Agoda.Analyzers.Test/Helpers/GenericTestHelpers/ConfigManager.cs
@@ -1,0 +1,23 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Agoda.Analyzers.Test.Helpers.GenericTestHelpers
+{
+    public static class ConfigManager
+    {
+        public static int[] GetLocationsFromConfig(string code)
+        {
+            return code.Substring(0, code.IndexOf(Environment.NewLine))
+                .Replace("/*", String.Empty)
+                .Replace("*/", String.Empty)
+                .Trim()
+                .Split(new[] { ',' }, StringSplitOptions.RemoveEmptyEntries)
+                .ToList()
+                .Select(entry => Convert.ToInt32(entry))
+                .ToArray();
+        }
+    }
+}
